requirements.txt
pandas
numpy
scikit-learn
joblib
watchdog
PyYAML
